fix(client): compile and lint the Android branch, and check it in CI
The Android branch is behind cfg, so a Linux host never compiles it — and the
attempt ladder that just landed there is real logic, not the no-ops the platform
layer had. Building for the target found three things no desktop build can:

- tarpc's Context::deadline is a std Instant, not a SystemTime, so the per-call
  deadline replacing tarpc's 10s default did not compile at all
- RunningInfo was never imported into rpc_server: an earlier edit did not match
  the actual import line and silently changed nothing
- an import that is live on desktop and dead on Android, which has to be scoped
  to the desktop path rather than removed

Then clippy for the target found sixteen more lints, none of them mine: the
Android sources have never been linted, because even the APK build only runs
cargo build.

So add a CI job that does it. cargo-ndk rather than a full APK build: it needs no
keystore and no Gradle, and it is stricter than what release.yml does. This is
the same gap the Windows job closed — a cfg-gated platform whose first compile
happened at release time.

+++ floppa-client/src-tauri/src/vpn/tunnel.rs
@@ -291,7 +291,7 @@ impl GotatunTunnel {
 
         // HACK: the `tun` crate stubs out MTU on Android (it just stores the value).
         // gotatun reads MTU from this, so we need to set it here with the correct value.
-        let mtu = config.get_mtu() as u16;
+        let mtu = config.get_mtu();
         tun_device
             .set_mtu(mtu)
             .map_err(|e| format!("Failed to set MTU: {}", e))?;
